# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VigiaApp API
  version: 1.1.0
  description: |
    API REST de VigiaApp. Parte de la especificación v1.0 (`VigiaApp/docs/especificacion_api_db.md`);
    las diferencias con esa versión están en `docs/CAMBIOS_CLIENTE.md`.

    - Todas las respuestas son `application/json; charset=utf-8` (RNF-04).
    - Claves en `snake_case`; los booleanos viajan como `true`/`false` (RNF-07).
    - Fechas con hora: ISO-8601 en UTC con `Z` y sin milisegundos (`2026-09-23T18:35:00Z`).
      Fechas de calendario: `AAAA-MM-DD`. Horas: `HH:MM`.
    - Todas las rutas, salvo `/auth/login`, `/auth/clave` y `/auth/recuperar`, exigen `Authorization: Bearer <token>`.
    - Cualquier respuesta que no sea 2xx trae el cuerpo `Error`.
servers:
  - url: https://{dominio}/api/v1
    variables:
      dominio:
        default: api.vigiaapp.example
  - url: http://localhost:3000/api/v1
    description: Desarrollo local
security:
  - bearer: []
...
paths:
...
  /rondines/semana:
    get:
      tags: [Rondines]
      summary: Grilla semanal (RF-07)
      description: |
        Los turnos entre `desde` y `desde + 6 días`. Por defecto `desde` es el lunes de la semana actual en la
        zona del barrio (`BARRIO_TZ`). Si un día no tiene turno cargado, no aparece.
      parameters:
        - { name: desde, in: query, required: false, schema: { type: string, format: date, example: '2026-09-21' } }
      responses:
        '200':
          description: Turnos ordenados por fecha
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Turno' } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
  /rondines/{id}/asistencia:
    post:
      tags: [Rondines]
      summary: '"Me presento" / "No me presento" (RF-08)'
      description: Con `presente:false` se avisa por correo a los miembros del comité para que reasignen el turno.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [presente]
              properties:
                presente: { type: boolean }
      responses:
        '200':
          description: El turno actualizado
          content: { application/json: { schema: { $ref: '#/components/schemas/Turno' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '404':
          description: El turno no existe o el vecino no está asignado.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /rondines:
    post:
      tags: [Comité]
      summary: Crear un turno (uno por fecha)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fecha, ruta]
              properties:
                fecha: { type: string, format: date }
                ruta: { type: string, maxLength: 200 }
                hora_inicio: { type: string, pattern: '^\d{2}:\d{2}$', default: '22:00' }
                hora_fin: { type: string, pattern: '^\d{2}:\d{2}$', default: '02:00' }
                vecino_ids: { type: array, items: { type: integer }, default: [] }
      responses:
        '201':
          description: Turno creado
          content: { application/json: { schema: { $ref: '#/components/schemas/Turno' } } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '409':
          description: Ya hay un turno en esa fecha.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /rondines/{id}:
    delete:
      tags: [Comité]
      summary: Borrar un turno
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Borrado junto con sus asignaciones. }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
  /rondines/{id}/asignaciones:
    get:
      tags: [Comité]
      summary: Quiénes hacen el turno y qué respondieron
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Asignaciones
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Asignacion' } }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
    put:
      tags: [Comité]
      summary: Reasignar el turno
      description: Reemplaza la lista. Quien sigue asignado conserva su respuesta; los nuevos quedan `pendiente`.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vecino_ids]
              properties:
                vecino_ids: { type: array, items: { type: integer } }
      responses:
        '200':
          description: Asignaciones resultantes
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Asignacion' } }
        '400': { $ref: '#/components/responses/DatosInvalidos' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '403': { $ref: '#/components/responses/Prohibido' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

  # ---------------------------------------------------------------- Alertas
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: 'Token de 64 hex que devuelven `/auth/login` y `/auth/clave`. Vale 24 h (RNF-03).'
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
...
    DatosInvalidos:
      description: Falta un campo o tiene formato incorrecto.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: datos_invalidos, mensaje: 'presente debe ser true o false.' } }
    NoAutorizado:
      description: Token ausente, inválido o vencido (o credenciales incorrectas en /auth/login).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_autorizado, mensaje: 'Sesión vencida. Volvé a iniciar sesión.' } }
    Prohibido:
      description: El vecino no tiene permiso (p. ej. no es del comité).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: prohibido, mensaje: 'Solo el comité puede hacer esto.' } }
    NoEncontrado:
      description: El recurso no existe.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { codigo: no_encontrado, mensaje: 'La alerta no existe.' } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [codigo, mensaje]
          properties:
            codigo:
              type: string
              enum: [datos_invalidos, no_autorizado, prohibido, no_encontrado, conflicto, demasiados_intentos, error_interno]
            mensaje: { type: string, description: 'Texto en castellano, apto para mostrar al usuario.' }
...
    Turno:
      type: object
      description: '`PatrolShift.fromJson` (lib/models/patrol_model.dart)'
      required: [id, fecha, hora_inicio, hora_fin, ruta, pareja, mi_estado]
      properties:
        id: { type: integer, example: 102 }
        fecha: { type: string, format: date, example: '2026-09-23' }
        hora_inicio: { type: string, example: '22:00' }
        hora_fin: { type: string, example: '02:00', description: 'Puede ser del día siguiente' }
        ruta: { type: string, example: Morelos → Juárez → Parque → Morelos }
        pareja: { type: array, items: { type: string }, example: [María R., Tomás R.] }
        mi_estado:
          type: [string, 'null']
          enum: [pendiente, confirmado, ausente, null]
          description: null si el vecino no está asignado a este turno
    Asignacion:
      type: object
      required: [vecino_id, nombre, estado, respondido_en]
      properties:
        vecino_id: { type: integer }
        nombre: { type: string, example: María Ríos }
        estado: { type: string, enum: [pendiente, confirmado, ausente] }
        respondido_en: { type: [string, 'null'], format: date-time }
...